implement compression logic
Keep track of both compressed and uncompressed byte counts.
Changed many functions and variables to use off_t instead of int when
dealing with file offsets and sizes.

dofork() currently always writes out a temp compressed file first,
rather than writing compressed bytes directly to the end of the archive.
This needs to be revisited to see if we can fwrite to a buffer and then
write that buffer to the archive in a loop.

Currently ENABLE_LZW_COMPRESSION is set to 0 as the archive produced
with it on isn't successfully being expanded. Continuing to debug this.